Remove plugin check
This exists check ensures that if the plugin has already loaded it won't
load again. This might be useful for plugins that shouldn't be loaded
more than once, but with a colorscheme, it means that we can't switch
to another colorscheme and come back to vim-monokai-tasty.

Obviously we want to be able to load the colorscheme anytime even if it
was loaded previously.